# optimize settings for realcam, allowing for changing codec