--- a/fakecam/fake.py
+++ b/fakecam/fake.py
@@ -29,23 +29,62 @@ def findFile(pattern, path):
                 return os.path.join(root, name)
     return None
 
+def get_codec_args_from_string(codec):
+    return (char for char in codec)
+
+
+def _log_camera_property_not_set(prop, value):
+    print("Cannot set camera property {} to {}. "
+          "Defaulting to auto-detected property set by opencv".format(prop, value))
+
+
 class RealCam:
-    def __init__(self, src, frame_width, frame_height, frame_rate):
+    def __init__(self, src, frame_width, frame_height, frame_rate, codec):
         self.cam = cv2.VideoCapture(src, cv2.CAP_V4L2)
         self.stopped = False
         self.frame = None
         self.lock = threading.Lock()
-        self._set_prop(cv2.CAP_PROP_FRAME_WIDTH, frame_width)
-        self._set_prop(cv2.CAP_PROP_FRAME_HEIGHT, frame_height)
-        self._set_prop(cv2.CAP_PROP_FPS, frame_rate)
+        self.get_camera_values("original")
+        c1, c2, c3, c4 = get_codec_args_from_string(codec)
+        self._set_codec(cv2.VideoWriter_fourcc(c1, c2, c3, c4))
+        self._set_frame_dimensions(frame_width, frame_height)
+        self._set_frame_rate(frame_rate)
+        self.get_camera_values("new")
 
-    def _set_prop(self, prop, value):
-        if self.cam.set(prop, value):
-            if value == self.cam.get(prop):
-                return True
+    def get_camera_values(self, status):
+        print(
+            "Real camera {} values are set as: {}x{} with {} FPS and video codec {}".format(
+                status,
+                self.get_frame_width(),
+                self.get_frame_height(),
+                self.get_frame_rate(),
+                self.get_codec()
+            )
+        )
 
-        print("Cannot set camera property {} to {}, used value: {}".format(prop, value, self.cam.get(prop)))
-        return False
+    def _set_codec(self, codec):
+        self.cam.set(cv2.CAP_PROP_FOURCC, codec)
+        if codec != self.get_codec():
+            _log_camera_property_not_set(cv2.CAP_PROP_FOURCC, codec)
+
+    def _set_frame_dimensions(self, width, height):
+        # width/height need to both be set before checking for any errors.
+        # If either are checked before setting both, either can be reported as not set properly
+        self.cam.set(cv2.CAP_PROP_FRAME_WIDTH, width)
+        self.cam.set(cv2.CAP_PROP_FRAME_HEIGHT, height)
+
+        if width != self.get_frame_width():
+            _log_camera_property_not_set(cv2.CAP_PROP_FRAME_WIDTH, width)
+        if height != self.get_frame_height():
+            _log_camera_property_not_set(cv2.CAP_PROP_FRAME_HEIGHT, height)
+
+    def _set_frame_rate(self, fps):
+        self.cam.set(cv2.CAP_PROP_FPS, fps)
+        if fps != self.get_frame_rate():
+            _log_camera_property_not_set(cv2.CAP_PROP_FPS, fps)
+
+    def get_codec(self):
+        return int(self.cam.get(cv2.CAP_PROP_FOURCC))
 
     def get_frame_width(self):
         return int(self.cam.get(cv2.CAP_PROP_FRAME_WIDTH))
@@ -85,6 +124,7 @@ class FakeCam:
         fps: int,
         width: int,
         height: int,
+        codec: str,
         scale_factor: float,
         no_background: bool,
         background_blur: int,
@@ -111,7 +151,7 @@ class FakeCam:
         self.foreground_image = foreground_image
         self.foreground_mask_image = foreground_mask_image
         self.scale_factor = scale_factor
-        self.real_cam = RealCam(webcam_path, width, height, fps)
+        self.real_cam = RealCam(webcam_path, width, height, fps, codec)
         # In case the real webcam does not support the requested mode.
         self.width = self.real_cam.get_frame_width()
         self.height = self.real_cam.get_frame_height()
@@ -340,6 +380,8 @@ def parse_args():
                         help="Set real webcam height")
     parser.add_argument("-F", "--fps", default=30, type=int,
                         help="Set real webcam FPS")
+    parser.add_argument("-C", "--codec", default='MJPG', type=str,
+                        help="Set real webcam codec")
     parser.add_argument("-S", "--scale-factor", default=0.5, type=float,
                         help="Scale factor of the image sent to BodyPix network")
     parser.add_argument("-B", "--bodypix-url", default="http://127.0.0.1:9000",
@@ -396,6 +438,7 @@ def main():
         fps=args.fps,
         width=args.width,
         height=args.height,
+        codec=args.codec,
         scale_factor=args.scale_factor,
         no_background=args.no_background,
         background_blur=getNextOddNumber(args.background_blur),