Proposed changes to optimize the real camera with the following changes:

1. Set default codec to `MJPG` to get better FPS at higher resolutions. The default detected by many cameras is `YUYV` which limits to 10 fps in many instances
2. Allow for codec to be set in the arg list if desired
3. setting both width and height before checking their validity If both are not set false errors occur
4. Presenting the configurable value before and after setting them

Various scenarios:

Using default settings set in parse_args
```bash
fakecam_1  | 2021-02-21T21:28:44.538579395Z Real camera original values are set as: 640x480 with 30 FPS and video codec 1448695129
fakecam_1  | 2021-02-21T21:28:44.605639190Z Real camera new values are set as: 1280x720 with 30 FPS and video codec 1196444237
fakecam_1  | 2021-02-21T21:28:44.606915970Z Running...
```

Setting the codec to `YUYV (-C=YUYV)`, resulting in fps that cannot be set to 30
```bash
fakecam_1  | 2021-02-21T21:24:40.992033721Z Real camera original values are set as: 640x480 with 30 FPS and video codec 1448695129
fakecam_1  | 2021-02-21T21:24:41.041994266Z Cannot set camera property 5 to 30. Defaulting to auto-detected property set by opencv
fakecam_1  | 2021-02-21T21:24:41.042025653Z Real camera new values are set as: 1280x720 with 10 FPS and video codec 1448695129
fakecam_1  | 2021-02-21T21:24:41.043258375Z Running...
```

Setting width to 1920 (which my camera does not support), resulting in width/height to be opencv auto-detected defaults allowed on my caemera with `MJPG` codec
```bash
fakecam_1  | 2021-02-21T21:28:01.093536631Z Real camera original values are set as: 640x480 with 30 FPS and video codec 1448695129
fakecam_1  | 2021-02-21T21:28:01.161294494Z Cannot set camera property 3 to 1920. Defaulting to auto-detected property set by opencv
fakecam_1  | 2021-02-21T21:28:01.161325520Z Real camera new values are set as: 1280x720 with 30 FPS and video codec 1196444237
fakecam_1  | 2021-02-21T21:28:01.162575545Z Running...
```